feat(codegen): BTreeMap for-in iteration + remove 13 i32 type fallbacks
- Implement Iterator[MapEntry[K,V]] for BTreeMapIter[K,V] (phase0p)
- Add MapEntry[K,V] type for key-value pairs yielded by iteration
- Fix type checker to substitute generics in Iterator::Item resolution
- Fix codegen to mangle generic type args in for-in iterator dispatch
- Fix codegen to request generic method instantiation for for-in loops
- Fix Maybe payload extraction via GEP instead of extractvalue (type mismatch)
- Remove 13 dangerous i32 fallbacks across AST/MIR codegen paths
- Move test executables to build/tests/ (out of bin/)
- Document Zig 0.16 incompatibility in README (use 0.15.2)
- Create phase0x task for remaining i32 fallback audit

+++ compiler/src/codegen/llvm/expr/closure.cpp
@@ -143,7 +143,7 @@ auto LLVMIRGen::gen_closure(const parser::ClosureExpr& closure) -> std::string {
     std::vector<std::string> param_names;
     for (size_t i = 0; i < closure.params.size(); ++i) {
         // Get parameter type: prefer explicit annotation, then inferred type from type checker
-        std::string param_type = "i32"; // default
+        std::string param_type = "i64"; // TML default integer type
         if (closure.params[i].second.has_value()) {
             param_type = llvm_type(*closure.params[i].second.value());
         } else if (i < closure.inferred_param_types.size() && closure.inferred_param_types[i]) {
